requirements.txt
annotated-types==0.7.0
pdfplumber==0.11.9
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.32.5
flask==3.1.3
sniffio==1.3.1
soupsieve==2.8.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
